py_assignment/py_assignment/requirements.txt
json
mysql-connector-python
flask